requirements/base.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
Pillow==10.4.0
django-filter==24.3
python-dateutil==2.9.0.post0
reportlab==4.2.5
python-dotenv==1.0.1

